[CORDA-1993]: Moved `fun logRawConfigurationErrors(...)` from `ValidateConfigurationCli` to `SharedNodeCmdLineOptions`.
